Refactoring packages in tests to align with their counterpart code
